requirements.txt
configargparse==1.2.3
h5py==2.10.0
keras==2.4.3
matplotlib==3.3.2
nibabel==3.1.1
numpy==1.19.1
opencv-python==4.4.0.42
owlready2==0.15
pandas==1.1.2
pydicom==2.0.0
scikit-image==0.17.2
scikit-learn==0.22.1
seaborn==0.11.0
tqdm==4.49.0
virtualenv==20.0.31
tensorflow-gpu==2.2.0